requirements.txt
django==1.8.2
psycopg2
djangorestframework
django-chart-tools
